requirements.txt
# Tests
mock>=0.8.0
Django>=1.4.3
django-nose
fabric
